Silence realloc() maximum size argument warning
gcc 9.3.1 emits the following warning:

  WDL/heapbuf.h:116:27: warning: argument 1 range [18446744071562067968, 18446744073709551615] exceeds maximum object size 9223372036854775807 [-Walloc-size-larger-than=]
    116 |         void *nbuf=realloc(m_buf, newalloc);
        |                    ~~~~~~~^~~~~~~~~~~~~~~~~~

This happens due to the implicit signed int to size_t conversion. An
unsigned int argument silences the issue.

+++ WDL/heapbuf.h
@@ -113,7 +113,7 @@ class WDL_HeapBuf
 
         if (newalloc != m_alloc || !m_buf)
         {
-        void *nbuf=realloc(m_buf,newalloc);
+        void *nbuf=realloc(m_buf, (unsigned)newalloc);
           if (!nbuf) 
           {
             if (!newalloc) return m_buf;
